# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Wildhaven Collectible Cards API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packs:
    get:
      operationId: listPacks
      tags: [packs]
      summary: List all card packs
      responses:
        "200":
          description: List of packs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pack"
    post:
      operationId: createPack
      tags: [packs]
      summary: Create a card pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackBody"
      responses:
        "201":
          description: Created pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pack"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /packs/{packId}:
    get:
      operationId: getPack
      tags: [packs]
      summary: Get a pack by ID with its cards
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pack with cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackWithCards"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePack
      tags: [packs]
      summary: Update a pack
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackBody"
      responses:
        "200":
          description: Updated pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pack"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePack
      tags: [packs]
      summary: Delete a pack
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /packs/{packId}/open:
    post:
      operationId: openPack
      tags: [packs]
      summary: Open a pack for a student (awards 3 cards)
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenPackBody"
      responses:
        "200":
          description: Cards awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenPackResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List cards, optionally filtered by pack
      parameters:
        - name: packId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: createCard
      tags: [cards]
      summary: Create a card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardBody"
      responses:
        "201":
          description: Created card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cards/{cardId}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get a card by ID
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCard
      tags: [cards]
      summary: Update a card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardBody"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Delete a card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Pack:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        cardBackImageUrl:
          type: ["string", "null"]
        commonChance:
          type: integer
        rareChance:
          type: integer
        epicChance:
          type: integer
        mythicChance:
          type: integer
        legendaryChance:
          type: integer
        color:
          type: ["string", "null"]
        cardCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - coverImageUrl
        - cardBackImageUrl
        - commonChance
        - rareChance
        - epicChance
        - mythicChance
        - legendaryChance
        - color
        - cardCount
        - createdAt
    PackWithCards:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        cardBackImageUrl:
          type: ["string", "null"]
        commonChance:
          type: integer
        rareChance:
          type: integer
        epicChance:
          type: integer
        mythicChance:
          type: integer
        legendaryChance:
          type: integer
        color:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
      required:
        - id
        - name
        - description
        - coverImageUrl
        - cardBackImageUrl
        - commonChance
        - rareChance
        - epicChance
        - mythicChance
        - legendaryChance
        - color
        - createdAt
        - cards
    CreatePackBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        cardBackImageUrl:
          type: ["string", "null"]
        commonChance:
          type: integer
        rareChance:
          type: integer
        epicChance:
          type: integer
        mythicChance:
          type: integer
        legendaryChance:
          type: integer
      required:
        - name
    UpdatePackBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        cardBackImageUrl:
          type: ["string", "null"]
        commonChance:
          type: integer
        rareChance:
          type: integer
        epicChance:
          type: integer
        mythicChance:
          type: integer
        legendaryChance:
          type: integer
        color:
          type: ["string", "null"]
    Card:
      type: object
      properties:
        id:
          type: integer
        packId:
          type: integer
        cardNumber:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        rarity:
          type: string
          enum: [Common, Rare, Epic, Mythic, Legendary]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - packId
        - cardNumber
        - name
        - imageUrl
        - rarity
        - createdAt
    CreateCardBody:
      type: object
      properties:
        packId:
          type: integer
        cardNumber:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        rarity:
          type: string
          enum: [Common, Rare, Epic, Mythic, Legendary]
      required:
        - packId
        - cardNumber
        - name
        - rarity
    UpdateCardBody:
      type: object
      properties:
        packId:
          type: integer
        cardNumber:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        rarity:
          type: string
          enum: [Common, Rare, Epic, Mythic, Legendary]
...
    OpenPackBody:
      type: object
      properties:
        studentId:
          type: integer
      required:
        - studentId
    OpenPackResult:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
      required:
        - cards
...