# lib/api-zod/src/generated/types/createCardBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wildhaven Collectible Cards API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCardBodyRarity } from "./createCardBodyRarity";

export interface CreateCardBody {
  packId: number;
  cardNumber: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  rarity: CreateCardBodyRarity;
}