# lib/api-zod/src/generated/types/updateCardBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wildhaven Collectible Cards API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCardBodyRarity } from "./updateCardBodyRarity";

export interface UpdateCardBody {
  packId?: number;
  cardNumber?: number;
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
  rarity?: UpdateCardBodyRarity;
}