# lib/api-zod/src/generated/types/updatePackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wildhaven Collectible Cards API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePackBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  cardBackImageUrl?: string | null;
  /** @nullable */
  openVideoUrl?: string | null;
  commonChance?: number;
  rareChance?: number;
  epicChance?: number;
  mythicChance?: number;
  legendaryChance?: number;
  /** @nullable */
  color?: string | null;
}